Init operation for cluster position map record was added

+++ core/src/main/java/com/orientechnologies/orient/core/storage/cache/OCacheEntry.java
@@ -40,7 +40,7 @@ public interface OCacheEntry {
 
   long getFileId();
 
-  long getPageIndex();
+  int getPageIndex();
 
   void acquireExclusiveLock();
