+++ core/src/main/java/com/orientechnologies/orient/core/storage/impl/local/paginated/atomicoperations/OCacheEntryChanges.java
@@ -54,7 +54,7 @@ public class OCacheEntryChanges implements OCacheEntry {
   }
 
   @Override
-  public long getPageIndex() {
+  public int getPageIndex() {
     return delegate.getPageIndex();
   }
