+++ core/src/main/java/com/orientechnologies/orient/core/storage/index/sbtree/multivalue/v2/OCellBTreeMultiValueV2.java
@@ -1347,7 +1347,7 @@ public final class OCellBTreeMultiValueV2<K> extends ODurableComponent implement
   private static UpdateBucketSearchResult addToTheRightRootBucket(final int keyIndex, final boolean splitLeaf,
       final int indexToSplit, final OCacheEntry rightBucketEntry, final ArrayList<Long> resultPath,
       final ArrayList<Integer> itemPointers) {
-    resultPath.add(rightBucketEntry.getPageIndex());
+    resultPath.add((long) rightBucketEntry.getPageIndex());
     itemPointers.add(0);
 
     if (splitLeaf) {
@@ -1366,7 +1366,7 @@ public final class OCellBTreeMultiValueV2<K> extends ODurableComponent implement
     itemPointers.add(-1);
     itemPointers.add(keyIndex);
 
-    resultPath.add(leftBucketEntry.getPageIndex());
+    resultPath.add((long) leftBucketEntry.getPageIndex());
     return new UpdateBucketSearchResult(itemPointers, resultPath, keyIndex);
   }
